updated InvoiceConsumer

+++ src/main/java/io/github/jonasfschuh/aws_RESTful/consumer/InvoiceConsumer.java
@@ -17,27 +17,26 @@ import javax.jms.JMSException;
 import javax.jms.TextMessage;
 import java.io.BufferedReader;
 import java.io.IOException;
+import java.io.InputStreamReader;
 
 @Service
 public class InvoiceConsumer {
-
     private static final Logger log = LoggerFactory.getLogger(
             InvoiceConsumer.class);
-    private final InvoiceRepository invoiceRepository;
-    private final ObjectMapper objectMapper;
-    private final AmazonS3 amazonS3;
+    private ObjectMapper objectMapper;
+    private InvoiceRepository invoiceRepository;
+    private AmazonS3 amazonS3;
 
     @Autowired
     public InvoiceConsumer(ObjectMapper objectMapper, InvoiceRepository invoiceRepository, AmazonS3 amazonS3) {
-        this.amazonS3 = amazonS3;
         this.objectMapper = objectMapper;
         this.invoiceRepository = invoiceRepository;
+        this.amazonS3 = amazonS3;
     }
 
     @JmsListener(destination = "${aws.sqs.queue.invoice.events.name}")
-    public void receiveS3Event(TextMessage textMessage) throws JMSException, IOException {
-
-        log.info("Invoice event received - Message: {}", textMessage);
+    public void receiveS3Event(TextMessage textMessage)
+            throws JMSException, IOException {
         SnsMessage snsMessage = objectMapper.readValue(textMessage.getText(),
                 SnsMessage.class);
 
@@ -45,42 +44,43 @@ public class InvoiceConsumer {
                 .readValue(snsMessage.getMessage(), S3EventNotification.class);
 
         processInvoiceNotification(s3EventNotification);
-
     }
 
-    private void processInvoiceNotification(S3EventNotification s3EventNotification) throws IOException {
+    private void processInvoiceNotification(S3EventNotification
+                                                    s3EventNotification) throws IOException {
+        if (s3EventNotification.getRecords() != null) {
+            for (S3EventNotification.S3EventNotificationRecord
+                    s3EventNotificationRecord : s3EventNotification.getRecords()) {
+                S3EventNotification.S3Entity s3Entity =
+                        s3EventNotificationRecord.getS3();
 
-        for (S3EventNotification.S3EventNotificationRecord
-             s3EventNotificationRecord : s3EventNotification.getRecords()) {
-            S3EventNotification.S3Entity s3Entity = s3EventNotificationRecord.getS3();
+                String bucketName = s3Entity.getBucket().getName();
+                String objectKey = s3Entity.getObject().getKey();
 
-            String bucketName = s3Entity.getBucket().getName();
-            String objectKey = s3Entity.getObject().getKey();
+                String invoiceFile = downloadObject(bucketName, objectKey);
 
-            String invoiceFile = downloadObject(bucketName, objectKey);
+                Invoice invoice = objectMapper.readValue(invoiceFile, Invoice.class);
+                log.info("Invoice received: {}", invoice.getInvoiceNumber());
 
-            Invoice invoice = objectMapper.readValue(invoiceFile, Invoice.class);
-            log.info("Invoice received: {} - Customer Name: {}",
-                    invoice.getInvoiceNumber(), invoice.getCustomerName());
+                invoiceRepository.save(invoice);
 
-            invoiceRepository.save(invoice);
-
-            amazonS3.deleteObject(bucketName, objectKey);
+                amazonS3.deleteObject(bucketName, objectKey);
+            }
         }
+
     }
 
-    private String downloadObject(String bucketName, String objectKey) throws IOException {
+    private String downloadObject(String bucketName, String objectKey)
+            throws IOException {
         S3Object s3Object = amazonS3.getObject(bucketName, objectKey);
 
         StringBuilder stringBuilder = new StringBuilder();
         BufferedReader bufferedReader = new BufferedReader(
-                new java.io.InputStreamReader(s3Object.getObjectContent()));
+                new InputStreamReader(s3Object.getObjectContent()));
         String content = null;
-        while ((content = bufferedReader.readLine())  != null) {
+        while ((content = bufferedReader.readLine()) != null) {
             stringBuilder.append(content);
         }
         return stringBuilder.toString();
     }
-
-
 }
